Límite de 5 productos, las alertas funcionan perfecto
