+++ client/src/redux/reducer.js
@@ -215,8 +215,8 @@ const rootReducer = (state = initialState, { type, payload }) => {
     
 
     case ADD_TO_CART_CANDY: {
-      if (state.cart.length >= 4) {
-        alert("No puedes seleccionar más de 6 productos.");
+      if (state.cart.length >= 5) {
+        alert("No puedes seleccionar más de 5 productos.");
         return state;
       }
